Rename project/sln back to UnwantedMethodCallsAnalyzer
Move analyzer readme to top level as this repo only plans to house this analyzer
