+++ .nuke
@@ -1,1 +1,1 @@
-source/RoslynAnalysers.sln
+source/UnwantedMethodCallsAnalyzer.sln
