+++ source/Tests/Tests.csproj
@@ -11,7 +11,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\RoslynAnalysers\RoslynAnalysers.csproj" />
+    <ProjectReference Include="..\UnwantedMethodCallsAnalyzer\UnwantedMethodCallsAnalyzer.csproj" />
 
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
     <PackageReference Include="nunit" Version="3.13.0" />
