+++ source/UnwantedMethodCallsAnalyzer.sln
@@ -3,13 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RoslynAnalysers", "RoslynAnalysers\RoslynAnalysers.csproj", "{2AA46A22-9D60-4737-8C2F-639B9636F872}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnwantedMethodCallsAnalyzer", "UnwantedMethodCallsAnalyzer\UnwantedMethodCallsAnalyzer.csproj", "{2AA46A22-9D60-4737-8C2F-639B9636F872}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{8F6A719A-C9EA-4836-80B9-A8FAD870031E}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_", "_", "{06CE9746-01A4-414B-BCDB-908B8B0BF84E}"
 	ProjectSection(SolutionItems) = preProject
-		RoslynAnalysers.sln.DotSettings = RoslynAnalysers.sln.DotSettings
+		UnwantedMethodCallsAnalyzer.sln.DotSettings = UnwantedMethodCallsAnalyzer.sln.DotSettings
 		..\.gitattributes = ..\.gitattributes
 		..\.gitignore = ..\.gitignore
 		..\.nuke = ..\.nuke
