+++ source/UnwantedMethodCallsAnalyzer.sln.DotSettings
