+++ source/UnwantedMethodCallsAnalyzer/tools/install.ps1
