+++ source/UnwantedMethodCallsAnalyzer/tools/uninstall.ps1
